pyproject.toml
[project]
name = "llm-mas"
version = "0.1.0"
description = "LLM Multi-Agent System architectures with LangChain and LangGraph"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "langchain>=0.3",
    "langchain-core>=0.3",
    "langchain-openai>=0.3",
    "langchain-community>=0.3",
    "langgraph>=0.3",
    "python-dotenv>=1.0",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ipykernel>=6.0",        # Jupyter support inside VS Code
    "ruff>=0.4",             # Linter + formatter
    "mypy>=1.10",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/llm_mas"]

# ── Ruff (linter + formatter) ──────────────────────────────────────────────
[tool.ruff]
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]   # pycodestyle, pyflakes, isort, pyupgrade

# ── Pytest ─────────────────────────────────────────────────────────────────
[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

# ── Mypy ───────────────────────────────────────────────────────────────────
[tool.mypy]
python_version = "3.12"
strict = false
ignore_missing_imports = true

[tool.uv]
default-groups = ["dev", "notebooks"]

[dependency-groups]
dev = [
    "ruff>=0.15.7",
]
notebooks = [
    "ipykernel>=6.0",
]
